Optimization, fix GetHoveredObjects

+++ src/SSGE_animation.c
@@ -17,7 +17,7 @@
  * \param draw The draw function of the animation if `type` is `SSGE_ANIM_FUNCTION`, ignored if not
  * \return The id of the animation
  */
-SSGEDECL uint32_t SSGE_Animation_Create(char *name, SSGE_AnimationType type, uint16_t frameCount, void (*draw)(SSGE_AnimationState *)) {
+SSGEDECL uint32_t SSGE_Animation_Create(char *name, SSGE_AnimationType type, uint32_t frameCount, void (*draw)(SSGE_AnimationState *)) {
     _assert_engine_init
 
     SSGE_Animation *anim = (SSGE_Animation *)malloc(sizeof(SSGE_Animation));
